Corrige função fila e retorna arquivo para pasta
